# Conflicts:
#	MealSync.Application/Common/Repositories/IAccountRepository.cs
#	MealSync.Application/UseCases/Accounts/Commands/CustomerLoginHandler.cs
#	MealSync.Application/UseCases/Accounts/Models/AccountResponse.cs
#	MealSync.Application/UseCases/Accounts/Models/LoginResponse.cs
#	MealSync.Domain/Enums/Roles.cs
#	MealSync.Infrastructure/Persistence/Contexts/MealSyncContext.cs
#	MealSync.Infrastructure/Persistence/Repositories/AccountRepository.cs

+++ MealSync.Application/UseCases/Test/Commands/TestValidateError/TestValidateErrorCommand.cs
@@ -6,8 +6,8 @@ namespace MealSync.Application.UseCases.Test.Commands.TestValidateError;
 
 public class TestValidateErrorCommand : ICommand<Result>
 {
-    public int Id { get; set; }
-    
+    public long Id { get; set; }
+
     public string Email { get; set; }
 }
